Move hook to its own file
