+++ components/brave_new_tab_ui/components/default/braveToday/cards/CardImage.tsx
@@ -5,7 +5,8 @@
 
 import * as React from 'react'
 import * as Card from '../cardSizes'
-import getBraveNewsController, * as BraveNews from '../../../../api/brave_news'
+import * as BraveNews from '../../../../api/brave_news'
+import { useUnpaddedImageUrl } from '../useUnpaddedImageUrl'
 
 type Props = {
   imageUrl?: string
@@ -14,60 +15,8 @@ type Props = {
   onLoaded?: () => any
 }
 
-const cache: { [url: string]: string } = {}
-
-export function useGetUnpaddedImage (paddedUrl: string | undefined, onLoaded?: () => any, useCache?: boolean) {
-  const [unpaddedUrl, setUnpaddedUrl] = React.useState('')
-
-  React.useEffect(() => {
-    const onReceiveUnpaddedUrl = (result: string) => {
-      if (useCache) cache[paddedUrl!] = result
-      setUnpaddedUrl(result)
-
-      if (onLoaded) window.requestAnimationFrame(() => onLoaded())
-    }
-
-    // Storybook method
-    // @ts-expect-error
-    if (window.braveStorybookUnpadUrl) {
-      // @ts-expect-error
-      window.braveStorybookUnpadUrl(paddedUrl)
-        .then(onReceiveUnpaddedUrl)
-      return
-    }
-
-    if (!paddedUrl) return
-
-    if (cache[paddedUrl]) {
-      onReceiveUnpaddedUrl(cache[paddedUrl])
-      return
-    }
-
-    let blobUrl: string
-    getBraveNewsController().getImageData({ url: paddedUrl })
-      .then(async (result) => {
-        if (!result.imageData) {
-          return
-        }
-
-        const blob = new Blob([new Uint8Array(result.imageData)], { type: 'image/*' })
-        blobUrl = URL.createObjectURL(blob)
-        onReceiveUnpaddedUrl(blobUrl)
-      })
-      .catch(err => {
-        console.error(`Error getting image for ${paddedUrl}.`, err)
-      })
-
-      // Only revoke the URL if we aren't using the cache.
-      return () => {
-        if (!useCache) URL.revokeObjectURL(blobUrl)
-      }
-  }, [paddedUrl])
-  return unpaddedUrl
-}
-
 export default function CardImage (props: Props) {
-  const unpaddedUrl = useGetUnpaddedImage(props.imageUrl, props.onLoaded)
+  const unpaddedUrl = useUnpaddedImageUrl(props.imageUrl, props.onLoaded)
   const [isImageLoaded, setIsImageLoaded] = React.useState(false)
   React.useEffect(() => {
     if (unpaddedUrl) {
